requirements.txt
wheel==0.33.4 
Django==2.2
Pillow==6.0.0
simplejson==3.16.0
djangorestframework==3.9.4
channels==2.2.0
channels-redis==2.4.0